Rename AdminSlice and AuthSlice
