+++ app/src/slices/admin/ActAdminAuth.tsx
@@ -2,7 +2,7 @@ import {createAsyncThunk} from '@reduxjs/toolkit'
 import axios from "axios";
 import {LOCAL_URL} from '../../constants/Config'
 import  { jwtDecode } from 'jwt-decode'
-import {authAdminState} from '../admin/authSlice'
+import {authAdminState} from './AuthSlice'
 
 export const loginAdmin = createAsyncThunk(
     '/admins/loginAdmin',async(adminCredential:authAdminState)=>{
