+++ app/src/slices/admin/authSlice.tsx
@@ -2,6 +2,7 @@ import {createSlice} from '@reduxjs/toolkit'
 import {loginAdmin} from './ActAdminAuth'
 
  export interface authAdminState{
+    authenticated:boolean |null,
     admin:{
         email:string | null,
         password: string | null
@@ -10,11 +11,12 @@ import {loginAdmin} from './ActAdminAuth'
     error:null|string,
 }
 const initialState : authAdminState ={
-    loading:false,
+    authenticated :false,
     admin:{
         email:  null ,
         password : null
     },
+    loading:false,
     error:null
 
 };  
@@ -32,12 +34,14 @@ const authAdminSlice = createSlice({
         .addCase(loginAdmin.fulfilled,(state,action)=>{
             state.loading=false;
             state.admin=action.payload;
+            state.authenticated=true;
             state.error=null;
         }) 
         .addCase(loginAdmin.rejected,(state,action)=>{
             state.loading=false;
             if (action.payload && typeof action.payload == "string") {
                 state.error = action.payload;
+                state.authenticated=false;
             }
            
            //state.error=action.error.message;
